requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Excel处理
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# AI服务
openai>=1.12.0
anthropic>=0.39.0

# 其他工具
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4

# 规则解析器依赖
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
python-docx==1.1.0
requests==2.31.0

# OCR支持（可选）
pytesseract==0.3.10
Pillow==10.1.0
easyocr==1.7.1

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
